Some work on the TIA class (I don't know how long the current code will be used, but at least we can improve it a little).
Eliminated the 'alignment' dimension in the various TIATables masks,
reducing the size of the arrays by a factor of 4.  I could never
figure out what alignment meant, until I looked at old TIA code in
the repo.  It seems that originally, there was an optimization
in the code to fill the array on 32-bit boundaries, instead of the
current 8-bit boundary.  As a result, the masks had to be defined
as 32-bits, or 4 groups of 8-bits.  Ah, that's where the 'alignment'
comes from.

Related to this, the colors and pointers in the TIA class are now
8-bit as well.  Essentially, the TIA class has been doing extra
work to align everything to 32-bit and never actually using the
results.  And it's been this way for 4+ years.

Older state files will no longer work, so the version # has been
bumped.