Some more code pruning.  Basically, all classname references to OpenGL have been changed to 'SDL2'.  From this point on, the OpenGL-specific code will be removed, and replaced by SDL2-specific hardware-accelerated code.
Cleaned up (mostly) the API of FrameBuffer vs FrameBufferSDL2.  Since the
FrameBuffer class is an emucore object, it should have no knowledge of SDL,
or any particular rendering toolkit for that matter (just in case we want
to move to SDL3 eventually :) ).

At this point, all checks for OpenGL have been removed, so if you don't
have it on your system, Stella will fail to compile and/or run.